requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

aiobotocore==2.13.1
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.3.0
appnope==0.1.4; platform_system == "Darwin"
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
botocore==1.34.131
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0; implementation_name == "pypy"
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6; sys_platform == "win32" or platform_system == "Windows"
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
executing==2.0.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
geojson==3.1.0
gitdb==4.0.11
gitpython==3.1.43
idna==3.7
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.2
jupyter-core==5.7.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
nest-asyncio==1.6.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0; sys_platform != "win32" and sys_platform != "emscripten"
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
prompt-toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0; sys_platform != "win32" and sys_platform != "emscripten"
pure-eval==0.2.2
pyarrow==16.1.0
pycparser==2.22; implementation_name == "pypy"
pydeck==0.9.1
pygments==2.18.0
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306; sys_platform == "win32" and platform_python_implementation != "PyPy"
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
s3fs==2024.6.1
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.36.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1; platform_system != "Darwin"
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
